Change objective function and inference procedure to match
The previous objective optimizes for a network which maps an audio segment and a pose tuple to a score proportional to the log likelihood of the sound having been emitted at any of the poses. The new objective decomposes this function into a combination of multiple functions of the audio snippet and a single pose. This allows for inference on single poses.